🔧 Avoid shipping it as new rule, as mutations have a nested input best practice
