+++ tests/test_validation_rules.py
@@ -2,6 +2,8 @@ import pytest
 from graphql import build_schema as schema
 
 from schemadiff.changes import Change, Criticality
+from schemadiff.changes.field import FieldArgumentAdded
+from schemadiff.changes.object import ObjectTypeFieldAdded
 from schemadiff.validation import evaluate_rules, ValidationResult, ValidationError
 from schemadiff.validation_rules import (
     ValidationRule,
@@ -234,6 +236,30 @@ def test_schema_added_field_no_desc():
 def test_cant_create_mutation_with_more_than_10_arguments():
     schema_restrictions = ['field-has-too-many-arguments']
 
+    class FieldHasTooManyArguments(ValidationRule):
+        """Restrict adding fields with too many top level arguments"""
+
+        name = "field-has-too-many-arguments"
+        limit = 10
+
+        def is_valid(self) -> bool:
+            if not isinstance(self.change, (ObjectTypeFieldAdded, FieldArgumentAdded)):
+                return True
+
+            if len(self.args) > self.limit:
+                return False
+            else:
+                return True
+
+        @property
+        def args(self):
+            return self.change.field.args or {}
+
+        @property
+        def message(self):
+            return f"Field `{self.change.parent.name}.{self.change.field_name}` has too many arguments " \
+                   f"({len(self.args)}>{self.limit}). Rule: {self.name}"
+
     old_schema = schema("""
     schema {
         mutation: Mutation
@@ -273,6 +299,30 @@ def test_cant_create_mutation_with_more_than_10_arguments():
 
 def test_cant_add_arguments_to_mutation_if_exceeds_10_args():
     schema_restrictions = ['field-has-too-many-arguments']
+
+    class FieldHasTooManyArguments(ValidationRule):
+        """Restrict adding fields with too many top level arguments"""
+
+        name = "field-has-too-many-arguments"
+        limit = 10
+
+        def is_valid(self) -> bool:
+            if not isinstance(self.change, (ObjectTypeFieldAdded, FieldArgumentAdded)):
+                return True
+
+            if len(self.args) > self.limit:
+                return False
+            else:
+                return True
+
+        @property
+        def args(self):
+            return self.change.field.args or {}
+
+        @property
+        def message(self):
+            return f"Field `{self.change.parent.name}.{self.change.field_name}` has too many arguments " \
+                   f"({len(self.args)}>{self.limit}). Rule: {self.name}"
 
     old_schema = schema("""
     schema {
